refactoring: move convert visitor to separate class
