+++ Source/LinqToDB/SqlQuery/SelectQueryOptimizer.cs
@@ -1014,7 +1014,7 @@ namespace LinqToDB.SqlQuery
 						var map = sql.Select.Columns.ToDictionary(c => c.Expression);
 						foreach (var condition in searchCondition)
 						{
-							var visitor = new QueryVisitor();
+							var visitor = new ConvertVisitor();
 							var newPredicate = visitor.Convert(condition.Predicate, e =>
 							{
 								if (e is ISqlExpression ex && map.TryGetValue(ex, out var newExpr))
